Fix documentation of `fidesops_meta` properties
As I was annotating some datasets I noticed that some of these docs seemed a bit different than the syntax used in example YAML files. I changed them to match what I saw there, but also not sure if perhaps it's just some new syntax that wasn't in the example files. Also happy to change any of the text to be more helpful, was just a quick pass at it.